Split a long rest go-file into short files for easy maintanence
