+++ api-runtime/rest-runtime/NicPublicIPRest.go
@@ -1,0 +1,273 @@
+// Cloud Control Manager's Rest Runtime of CB-Spider.
+// The CB-Spider is a sub-Framework of the Cloud-Barista Multi-Cloud Project.
+// The CB-Spider Mission is to connect all the clouds with a single interface.
+//
+//      * Cloud-Barista: https://github.com/cloud-barista
+//
+// by CB-Spider Team, 2019.
+
+package restruntime
+
+
+/****************************
+//================ VNic Handler
+func createVNic(c echo.Context) error {
+	cblog.Info("call createVNic()")
+
+        var req struct {
+                ConnectionName string
+                ReqInfo cres.VNicReqInfo
+        }
+
+        if err := c.Bind(&req); err != nil {
+                return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+        }
+
+        cldConn, err := ccm.GetCloudConnection(req.ConnectionName)
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	handler, err := cldConn.CreateVNicHandler()
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	info, err := handler.CreateVNic(req.ReqInfo)
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	return c.JSON(http.StatusOK, &info)
+}
+
+func listVNic(c echo.Context) error {
+	cblog.Info("call listVNic()")
+
+        var req struct {
+                ConnectionName string
+        }
+
+        if err := c.Bind(&req); err != nil {
+                return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+        }
+
+        cldConn, err := ccm.GetCloudConnection(req.ConnectionName)
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	handler, err := cldConn.CreateVNicHandler()
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	infoList, err := handler.ListVNic()
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+        var jsonResult struct {
+                Result []*cres.VNicInfo `json:"vnic"`
+        }
+        if infoList == nil {
+                infoList = []*cres.VNicInfo{}
+        }
+        jsonResult.Result = infoList
+        return c.JSON(http.StatusOK, &jsonResult)
+}
+
+func getVNic(c echo.Context) error {
+	cblog.Info("call getVNic()")
+
+        var req struct {
+                ConnectionName string
+        }
+
+        if err := c.Bind(&req); err != nil {
+                return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+        }
+
+        cldConn, err := ccm.GetCloudConnection(req.ConnectionName)
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	handler, err := cldConn.CreateVNicHandler()
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	info, err := handler.GetVNic(c.Param("VNicId"))
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	return c.JSON(http.StatusOK, &info)
+}
+
+func deleteVNic(c echo.Context) error {
+	cblog.Info("call deleteVNic()")
+
+        var req struct {
+                ConnectionName string
+        }
+
+        if err := c.Bind(&req); err != nil {
+                return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+        }
+
+        cldConn, err := ccm.GetCloudConnection(req.ConnectionName)
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	handler, err := cldConn.CreateVNicHandler()
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	result, err := handler.DeleteVNic(c.Param("VNicId"))
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+        resultInfo := BooleanInfo{
+                Result: strconv.FormatBool(result),
+        }
+
+	return c.JSON(http.StatusOK, &resultInfo)
+}
+
+//================ PublicIP Handler
+func createPublicIP(c echo.Context) error {
+	cblog.Info("call createPublicIP()")
+
+        var req struct {
+                ConnectionName string
+                ReqInfo cres.PublicIPReqInfo
+        }
+
+        if err := c.Bind(&req); err != nil {
+                return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+        }
+
+        cldConn, err := ccm.GetCloudConnection(req.ConnectionName)
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	handler, err := cldConn.CreatePublicIPHandler()
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	info, err := handler.CreatePublicIP(req.ReqInfo)
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	return c.JSON(http.StatusOK, &info)
+}
+
+func listPublicIP(c echo.Context) error {
+	cblog.Info("call listPublicIP()")
+
+        var req struct {
+                ConnectionName string
+        }
+
+        if err := c.Bind(&req); err != nil {
+                return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+        }
+
+        cldConn, err := ccm.GetCloudConnection(req.ConnectionName)
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	handler, err := cldConn.CreatePublicIPHandler()
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	infoList, err := handler.ListPublicIP()
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+        var jsonResult struct {
+                Result []*cres.PublicIPInfo `json:"publicip"`
+        }
+        if infoList == nil {
+                infoList = []*cres.PublicIPInfo{}
+        }
+        jsonResult.Result = infoList
+        return c.JSON(http.StatusOK, &jsonResult)
+}
+
+func getPublicIP(c echo.Context) error {
+	cblog.Info("call getPublicIP()")
+
+        var req struct {
+                ConnectionName string
+        }
+
+        if err := c.Bind(&req); err != nil {
+                return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+        }
+
+        cldConn, err := ccm.GetCloudConnection(req.ConnectionName)
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	handler, err := cldConn.CreatePublicIPHandler()
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	info, err := handler.GetPublicIP(c.Param("PublicIPId"))
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	return c.JSON(http.StatusOK, &info)
+}
+
+func deletePublicIP(c echo.Context) error {
+	cblog.Info("call deletePublicIP()")
+
+        var req struct {
+                ConnectionName string
+        }
+
+        if err := c.Bind(&req); err != nil {
+                return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+        }
+
+        cldConn, err := ccm.GetCloudConnection(req.ConnectionName)
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	handler, err := cldConn.CreatePublicIPHandler()
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+	result, err := handler.DeletePublicIP(c.Param("PublicIPId"))
+	if err != nil {
+		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
+	}
+
+        resultInfo := BooleanInfo{
+                Result: strconv.FormatBool(result),
+        }
+
+	return c.JSON(http.StatusOK, &resultInfo)
+}
+****************************/
+
+
+
